Rename repository from `swirl` to `swirlc`
This commit renames the whole repository from `swirl` to `swirlc`, which
stands for SWIRL Compiler. The main reason is that the `swirl` package
is already present on PyPI.
