+++ .flake8
@@ -1,4 +1,4 @@
 [flake8]
-exclude = swirl/antlr
+exclude = swirlc/antlr
 max-line-length = 88
 extend-ignore = E203,E501
