+++ .github/workflows/ci-tests.yaml
@@ -11,7 +11,7 @@ concurrency:
   cancel-in-progress: true
 jobs:
   code-ql-check:
-    name: "SWIRL CodeQL check"
+    name: "SWIRL Compiler CodeQL check"
     runs-on: ubuntu-22.04
     permissions:
       security-events: write
@@ -23,7 +23,7 @@ jobs:
           languages: python
       - uses: github/codeql-action/analyze@v3
   docker-image:
-    name: "SWIRL Docker image tests"
+    name: "SWIRL Compiler Docker image tests"
     runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v4
@@ -33,9 +33,9 @@ jobs:
         uses: docker/build-push-action@v5
         with:
           load: true
-          tags: alphaunito/swirl:latest
+          tags: alphaunito/swirlc:latest
   static-checks:
-    name: "SWIRL static checks"
+    name: "SWIRL Compiler static checks"
     runs-on: ubuntu-22.04
     strategy:
       matrix:
@@ -52,9 +52,9 @@ jobs:
             requirements.txt
             test-requirements.txt
             tox.ini
-      - name: "Install Python Dependencies and StreamFlow"
+      - name: "Install Python Dependencies and SWIRL Compiler"
         run: |
           python -m pip install tox --user
           python -m pip install . --user
-      - name: "Run SWIRL static analysis via Tox"
+      - name: "Run SWIRL Compiler static analysis via Tox"
         run: tox
