+++ .github/workflows/release.yaml
@@ -20,18 +20,18 @@ jobs:
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
-      - name: "Get SWIRL version"
-        run: echo "SWIRL_VERSION=$(cat swirl/version.py | grep -oP '(?<=VERSION = \")(.*)(?=\")')" >> $GITHUB_ENV
+      - name: "Get SWIRL Compiler version"
+        run: echo "SWIRLC_VERSION=$(cat swirlc/version.py | grep -oP '(?<=VERSION = \")(.*)(?=\")')" >> $GITHUB_ENV
       - name: "Check if Docker image already exists"
-        run: echo "NEW_IMAGE=$(docker buildx imagetools inspect alphaunito/swirl:${SWIRL_VERSION} > /dev/null 2>&1; echo $?)" >> $GITHUB_ENV
+        run: echo "NEW_IMAGE=$(docker buildx imagetools inspect alphaunito/swirlc:${SWIRLC_VERSION} > /dev/null 2>&1; echo $?)" >> $GITHUB_ENV
       - name: "Build Docker image"
         if: ${{ env.NEW_IMAGE == 1 }}
         uses: docker/build-push-action@v5
         with:
           push: true
           tags: |
-            alphaunito/swirl:${{ env.SWIRL_VERSION }}
-            alphaunito/swirl:latest
+            alphaunito/swirlc:${{ env.SWIRLC_VERSION }}
+            alphaunito/swirlc:latest
   github:
     name: "Create GitHub Release"
     runs-on: ubuntu-22.04
@@ -40,13 +40,13 @@ jobs:
     if: ${{ github.event.workflow_run.conclusion == 'success' }}
     steps:
       - uses: actions/checkout@v4
-      - name: "Get SWIRL version"
-        run: echo "SWIRL_VERSION=$(cat swirl/version.py | grep -oP '(?<=VERSION = \")(.*)(?=\")')" >> $GITHUB_ENV
+      - name: "Get SWIRL Compiler version"
+        run: echo "SWIRLC_VERSION=$(cat swirlc/version.py | grep -oP '(?<=VERSION = \")(.*)(?=\")')" >> $GITHUB_ENV
       - name: "Check tag existence"
         uses: mukunku/tag-exists-action@v1.6.0
         id: check-tag
         with:
-          tag: ${{ env.SWIRL_VERSION }}
+          tag: ${{ env.SWIRLC_VERSION }}
       - name: "Create Release"
         id: create-release
         uses: actions/create-release@v1
@@ -54,8 +54,8 @@ jobs:
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         with:
-          tag_name: ${{ env.SWIRL_VERSION }}
-          release_name: ${{ env.SWIRL_VERSION }}
+          tag_name: ${{ env.SWIRLC_VERSION }}
+          release_name: ${{ env.SWIRLC_VERSION }}
           draft: false
           prerelease: false
   pypi:
@@ -63,7 +63,7 @@ jobs:
     runs-on: ubuntu-22.04
     environment:
       name: pypi
-      url: https://pypi.org/project/swirl
+      url: https://pypi.org/project/swirlc
     permissions:
       id-token: write
     if: ${{ github.event.workflow_run.conclusion == 'success' }}
@@ -72,15 +72,15 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: "3.10"
-      - name: "Get SWIRL version"
-        run: echo "SWIRL_VERSION=$(cat swirl/version.py | grep -oP '(?<=VERSION = \")(.*)(?=\")')" >> $GITHUB_ENV
+      - name: "Get SWIRL Compiler version"
+        run: echo "SWIRLC_VERSION=$(cat swirlc/version.py | grep -oP '(?<=VERSION = \")(.*)(?=\")')" >> $GITHUB_ENV
       - name: "Get PyPI version"
-        run: echo "PYPI_VERSION=$(pip index versions --pre swirl | grep swirl | sed 's/.*(\(.*\))/\1/')" >> $GITHUB_ENV
+        run: echo "PYPI_VERSION=$(pip index versions --pre swirlc | grep swirlc | sed 's/.*(\(.*\))/\1/')" >> $GITHUB_ENV
       - name: "Build Python packages"
-        if: ${{ env.SWIRL_VERSION != env.PYPI_VERSION }}
+        if: ${{ env.SWIRLC_VERSION != env.PYPI_VERSION }}
         run: |
           python -m pip install build --user
           python -m build --sdist --wheel --outdir dist/ .
       - name: "Publish package to PyPI"
         uses: pypa/gh-action-pypi-publish@release/v1
-        if: ${{ env.SWIRL_VERSION != env.PYPI_VERSION }}
+        if: ${{ env.SWIRLC_VERSION != env.PYPI_VERSION }}
