+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.12-alpine3.20 AS builder
 
-ENV VIRTUAL_ENV="/opt/swirl"
+ENV VIRTUAL_ENV="/opt/swirlc"
 ENV PATH="${VIRTUAL_ENV}/bin:${PATH}"
 
 COPY ./pyproject.toml ./MANIFEST.in ./LICENSE ./README.md /build/
@@ -9,7 +9,7 @@ COPY ./requirements.txt           \
      ./lint-requirements.txt      \
      ./test-requirements.txt      \
      /build/
-COPY ./swirl /build/swirl
+COPY swirlc /build/swirlc
 
 RUN cd build \
     && python -m venv ${VIRTUAL_ENV} \
@@ -21,7 +21,7 @@ LABEL maintainer="Iacopo Colonnelli <iacopo.colonnelli@unito.it>"
 LABEL maintainer="Doriana Medić <doriana.medic@unito.it>"
 LABEL maintainer="Alberto Mulone <alberto.mulone@unito.it>"
 
-ENV VIRTUAL_ENV="/opt/swirl"
+ENV VIRTUAL_ENV="/opt/swirlc"
 ENV PATH="${VIRTUAL_ENV}/bin:${PATH}"
 
 COPY --from=builder ${VIRTUAL_ENV} ${VIRTUAL_ENV}
