+++ Makefile
@@ -1,8 +1,8 @@
 codespell:
-	codespell -w $(shell git ls-files | grep -v swirl/antlr)
+	codespell -w $(shell git ls-files | grep -v swirlc/antlr)
 
 codespell-check:
-	codespell $(shell git ls-files | grep -v swirl/antlr)
+	codespell $(shell git ls-files | grep -v swirlc/antlr)
 
 coverage.xml: testcov
 	coverage xml
@@ -11,16 +11,16 @@ coverage-report: testcov
 	coverage report
 
 flake8:
-	flake8 --exclude swirl/antlr swirl tests
+	flake8 --exclude swirlc/antlr swirlc tests
 
 format:
-	black swirl tests
+	black swirlc tests
 
 format-check:
-	black --diff --check swirl tests
+	black --diff --check swirlc tests
 
 pyupgrade:
-	pyupgrade --py3-only --py38-plus $(shell git ls-files | grep .py | grep -v swirl/antlr)
+	pyupgrade --py3-only --py38-plus $(shell git ls-files | grep .py | grep -v swirlc/antlr)
 
 test:
 	python -m pytest -rs ${PYTEST_EXTRA}
