+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "swirl"
+name = "swirlc"
 authors = [
     {name = "Iacopo Colonnelli", email = "iacopo.colonnelli@unito.it"},
     {name = "Doriana Medić", email = "doriana.medic@unito.it"},
@@ -31,30 +31,30 @@ classifiers = [
 dynamic = ["dependencies", "optional-dependencies", "version"]
 
 [project.scripts]
-swirl = "swirl.main:run"
+swirlc = "swirlc.main:run"
 
 [project.urls]
-Package = "https://pypi.org/project/swirl"
-Repository = "https://github.com/alpha-unito/swirl"
-Docker = "https://hub.docker.com/r/alphaunito/swirl"
+Package = "https://pypi.org/project/swirlc"
+Repository = "https://github.com/alpha-unito/swirlc"
+Docker = "https://hub.docker.com/r/alphaunito/swirlc"
 
 [tool.setuptools]
 packages = [
-    "swirl",
-    "swirl.antlr",
-    "swirl.compiler",
-    "swirl.config",
-    "swirl.core",
-    "swirl.translator"
+    "swirlc",
+    "swirlc.antlr",
+    "swirlc.compiler",
+    "swirlc.config",
+    "swirlc.core",
+    "swirlc.translator"
 ]
 zip-safe = true
 
 [tool.setuptools.package-data]
-"swirl.config" = ["schemas/v1.0/*.json"]
+"swirlc.config" = ["schemas/v1.0/*.json"]
 
 [tool.setuptools.dynamic]
 dependencies = {file = "requirements.txt"}
-version = {attr = "swirl.version.VERSION"}
+version = {attr = "swirlc.version.VERSION"}
 
 [tool.setuptools.dynamic.optional-dependencies]
 bandit = {file = "bandit-requirements.txt"}
@@ -63,8 +63,8 @@ test = {file = "test-requirements.txt"}
 
 [tool.coverage.run]
 branch = true
-source_pkgs = ["swirl"]
-omit = ["swirl/antlr/*"]
+source_pkgs = ["swirlc"]
+omit = ["swirlc/antlr/*"]
 
 [tool.coverage.report]
 exclude_lines = [
@@ -81,9 +81,9 @@ exclude_lines = [
 ]
 ignore_errors = true
 omit = [
-    "swirl/__main__.py",
+    "swirlc/__main__.py",
     "tests/*"
 ]
 
 [tool.black]
-exclude = "swirl/antlr"
+exclude = "swirlc/antlr"
