--- swirl/__main__.py
@@ -1,7 +1,0 @@
-"""Default entrypoint for the swirl module."""
-
-import sys
-
-from swirl import main
-
-main.main(sys.argv[1:])
