--- swirl/compiler/__init__.py
@@ -1,3 +1,0 @@
-from swirl.compiler.default import DefaultTarget
-
-targets = {"default": DefaultTarget}
