--- swirl/translator/__init__.py
@@ -1,3 +1,0 @@
-from swirl.translator.dax_translator import DAXTranslator
-
-translator_classes = {"dax": DAXTranslator}
