+++ swirlc/__main__.py
@@ -1,0 +1,7 @@
+"""Default entrypoint for the swirlc module."""
+
+import sys
+
+from swirlc import main
+
+main.main(sys.argv[1:])
