+++ swirlc/antlr/SWIRLListener.py
@@ -1,4 +1,4 @@
-# Generated from /swirl/grammar/SWIRL.g4 by ANTLR 4.13.1
+# Generated from /swirlc/grammar/SWIRL.g4 by ANTLR 4.13.1
 from antlr4 import *
 if "." in __name__:
     from .SWIRLParser import SWIRLParser
