+++ swirlc/compiler/__init__.py
@@ -1,0 +1,3 @@
+from swirlc.compiler.default import DefaultTarget
+
+targets = {"default": DefaultTarget}
