+++ swirlc/config/schemas/v1.0/config_schema.json
