+++ swirlc/translator/__init__.py
@@ -1,0 +1,3 @@
+from swirlc.translator.dax_translator import DAXTranslator
+
+translator_classes = {"dax": DAXTranslator}
