+++ tox.ini
@@ -26,7 +26,7 @@ setenv =
   py3.{8,9,10,11,12}-unit: LC_ALL = C.UTF-8
 
 [testenv:bandit]
-commands = bandit -r swirl
+commands = bandit -r swirlc
 deps =
   -rrequirements.txt
   -rbandit-requirements.txt
